IFDP/requirements.txt
async-timeout==4.0.2
azure-core==1.22.1
azure-cosmos==4.2.0
Brotli==1.0.9
bson==0.5.8
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dash==2.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
Flask==2.1.2
Flask-Caching==1.10.1
Flask-Compress==1.12
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
plotly==5.7.0
pymongo==3.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-memcached==1.59
pytz==2022.1
redis==4.3.1
requests==2.27.1
tenacity==8.0.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.8.0
